requirements.txt
astronomer-cosmos==1.4.3
dbt-core==1.8.1
dbt-athena-community==1.8.2